requirements.txt
shiny 
shinyswatch 
pandas 
openpyxl 
jinja2 
matplotlib 
seaborn 
plotnine
shinywidgets 
plotly 
holoviews 
panel 
hvplot 
ipyleaflet
jupyter_bokeh 